Add listener for account change event. Handle network switches without full reload.
