+++ src/App.scss
@@ -183,6 +183,12 @@
     }
   }
 
+  .wallet-button {
+    button {
+      font-size: 1rem;
+    }
+  }
+
   .modal-content {
     button {
       width: 33rem !important;
