+++ src/App.tsx
@@ -8,7 +8,8 @@ import HamburgerMenu from './components/hamburger-menu/hamburger-menu';
 import { Character, DefaultCharacter } from './types/Character';
 import { Contract, Signer, Web3Provider } from 'zksync-web3';
 import { ZkEraNft } from './types/ZkEraNFT';
-import Web3Context from './context/web3-context';
+import Web3Context, { Web3ContextType } from './context/web3-context';
+import { initContracts } from './utils/web3-helper';
 
 function App() {
   const LocalStorageCharacterKey = 'mnm-character-v1';
@@ -26,25 +27,50 @@ function App() {
   const [characterIDs, setCharacterIDs] = useState<string[]>([]);
   const [walletAddress, setWalletAddress] = useState<string>("");
 
+  const web3Context: Web3ContextType = {
+    provider,
+    setProvider,
+    signer,
+    setSigner,
+    charactersContractInstance,
+    setCharactersContractInstance,
+    nfts,
+    setNfts,
+    characterIDs,
+    setCharacterIDs,
+    walletAddress,
+    setWalletAddress,
+  };
+
   useEffect(() => {
     localStorage.setItem(LocalStorageCharacterKey, JSON.stringify(character));
   }, [character]);
 
+  useEffect(() => {
+    // Listen for accountsChanged event
+    (window as any).ethereum.on("accountsChanged", async (accounts: any) => {
+      const provider = new Web3Provider((window as any).ethereum);
+      setProvider(provider);
+
+      const signerInstance = provider.getSigner();
+      setSigner(signerInstance);
+
+      setWalletAddress(accounts[0]);
+
+      await initContracts(web3Context, setCharacter, provider, signerInstance);
+    });
+  }, [
+    setProvider,
+    setSigner,
+    setCharacter,
+    setWalletAddress,
+    setNfts,
+    setCharactersContractInstance,
+    setCharacterIDs,
+  ]);
+
   return (
-    <Web3Context.Provider value={{
-      provider,
-      setProvider,
-      signer,
-      setSigner,
-      charactersContractInstance,
-      setCharactersContractInstance,
-      nfts,
-      setNfts,
-      characterIDs,
-      setCharacterIDs,
-      walletAddress,
-      setWalletAddress,
-    }}>
+    <Web3Context.Provider value={web3Context}>
       <div className="App">
         <div className="background-image"></div>
         <div className="background-filter"></div>
