+++ src/components/wallet-button/wallet-button.tsx
@@ -1,19 +1,15 @@
 import { useState, useEffect, useContext, Dispatch, SetStateAction } from "react";
-import { Web3Provider, Signer, Contract } from "zksync-web3";
+import { Web3Provider } from "zksync-web3";
 import Web3Context from "../../context/web3-context";
 import {
-  CHARACTERS_ADDRESS,
-  CHARACTERS_CONTRACT_ABI,
-  NFT_CONTRACT_ADDRESS,
-  NFT_CONTRACT_ABI,
   NETWORK_NAME,
   NETWORK_ID,
 } from "../../constants/constants";
-import { ZkEraNft } from "../../types/ZkEraNFT";
 import { Address } from "zksync-web3/build/src/types";
 import { Character } from "../../types/Character";
-import { CharacterSmartContract } from "../../types/CharacterSmartContract";
 import "./wallet-button.scss";
+import { initContracts } from "../../utils/web3-helper";
+import { ethers } from "ethers";
 
 function WalletButton({
     character,
@@ -35,92 +31,37 @@ function WalletButton({
     return `${start}...${end}`;
   };
 
-  const initContracts = async (provider: Web3Provider, signer: Signer) => {
-    if (provider && signer) {
-      const charactersContract = new Contract(
-        CHARACTERS_ADDRESS,
-        CHARACTERS_CONTRACT_ABI,
-        signer
-      );
-
-      web3Context.setCharactersContractInstance(charactersContract);
-
-      const allCharacterIDs: string[] = await charactersContract.getCharacterIDs();
-      console.log(`getCharacterIDs: ${allCharacterIDs.join(",")}`);
-
-      if (allCharacterIDs.length > 0) {
-        web3Context.setCharacterIDs(allCharacterIDs);
-        const characterData: CharacterSmartContract = await charactersContract.getCharacter(allCharacterIDs[0]);
-        console.log(`Current character from smart contract: ${characterData}`);
-
-        const character = JSON.parse(characterData.jsonData) as Character;
-        
-        setCharacter(character);
-    } else {
-        web3Context.setCharacterIDs(["first-character"]);
-      }
-
-      if (allCharacterIDs.length == 0) {
-        return;
-      }
-      
-      const nftContract = new Contract(
-        NFT_CONTRACT_ADDRESS,
-        NFT_CONTRACT_ABI,
-        signer
-      );
-
-      const address = await signer.getAddress();
-      const balance = 0 // TODO: await nftContract.balanceOf(address);
-      if (balance > 0) {
-        let ownedNfts: ZkEraNft[] = [];
-        const ownedTokensResponse = await nftContract.tokensOfOwner(address);
-
-        for (let i = 0; i < ownedTokensResponse.length; i++) {
-          const tokenId = ownedTokensResponse[i];
-
-          const tokenURI = await nftContract.tokenURI(tokenId);
-          if (tokenURI == undefined || tokenURI == "") {
-            continue;
-          }
-
-          const response = await fetch(tokenURI);
-          if (!response.ok) {
-            continue;
-          }
-
-          ownedNfts.push((await response.json()) as ZkEraNft);
-        }
-
-        web3Context.setNfts(ownedNfts);
-      } else {
-        web3Context.setNfts([]);
-      }
-    }
-  };
-
   const checkNetwork = async () => {
     if ((window as any).ethereum) {
       const currentChainId = await (window as any).ethereum.request({
         method: "eth_chainId",
       });
 
-      if (currentChainId == NETWORK_ID) setNetworkOk(true);
+      if (currentChainId == NETWORK_ID) {
+        setNetworkOk(true);
+      }
     }
   };
 
   const switchNetwork = async () => {
-    await (window as any).ethereum.request({
-      method: "wallet_switchEthereumChain",
-      params: [{ chainId: NETWORK_ID }],
-    });
-
-    // refresh
-    window.location.reload();
+    try {
+      await (window as any).ethereum.request({
+        method: "wallet_switchEthereumChain",
+        params: [{ chainId: ethers.utils.hexlify(NETWORK_ID).replace("0x0", "0x") }],
+      });
+  
+      setNetworkOk(true);
+    } catch (error) {
+      // User did not update networks
+      console.error(error); 
+    }
   };
 
   const connectWallet = async () => {
-    if (!networkOk) await switchNetwork();
+    if (!networkOk) {
+      await switchNetwork();
+    }
+
     try {
       if ((window as any).ethereum) {
         const provider = new Web3Provider((window as any).ethereum);
@@ -133,7 +74,7 @@ function WalletButton({
 
         web3Context.setWalletAddress(data[0]);
 
-        await initContracts(provider, signerInstance);
+        await initContracts(web3Context, setCharacter, provider, signerInstance);
       }
     } catch (error) {
       console.error("Error connecting DApp to your wallet");
@@ -145,9 +86,9 @@ function WalletButton({
     <div className="wallet-button">
       {!networkOk ? (
         <button
-          onClick={switchNetwork}
+          onClick={connectWallet}
         >
-          Wrong network. Switch to {NETWORK_NAME}
+          Wrong network, switch to {NETWORK_NAME}
         </button>
       ) : (
         <button
